Added some CFR+ training code.
The classes added are a trainer class, node class, and information set class. I
was debating whether to represent the information set as a dict or a class, but
I'm going with class because it contains a lot of information, and will also be
useful for Texas Hold'em. It is unclear how much bounds checking I should
really do on the inputs to an InfoSet, since it may be too expensive and hard
to hit all edge cases. I have the bounds checking in for now but may delete later.