Adapted code for the turn abstraction.
The turn abstraction will bucket together all possible six-card hands.
The good news on the turn is that there are only 44 river rollouts to
compute for each opponent hand, so computing equity distributions is
easier. The bad news is that there are far more archetypal hands, so the
memory footprint of the abstraction is much larger. I don't yet know
exactly how many archetypal turn hands there are.