Local best response calculation against a chump opponent.
I completed implementing a Local Best Response (LBR) algorithm to
calculate a lower bound of the exploitability of an opponent, given
their strategy. This will be useful for evaluating the blueprint
strategy and the overall quality of the bot. This code now plays against
a chump opponent that always check-calls, which in theory is exploitable
to 50 BB/h with 200 BB stacks. My code gets ~45 BB/h exploitability,
which suggests that it is working properly for this case. The next step
will be testing other chump opponents and then the real blueprint
strategy.

This Python implementation can generate a reasonably accurate
exploitability estimate in under 10 minutes. The limiting factor is the
number of opponent hand samples for rollouts, which can be reduced to a
lower number (~50) without too much increased variance.