Evaluated the exploiter against the random-action chump.
There might be bugs in this and the exploitability doesn't exactly match
that of the paper, but my distribution over bet sizes must be different
then what they did. They probably had the chump choose all-in with 1/4
probability or something like that. When I biased the chump towards
larger bets, the exploitability was much higher, suggesting that things
are working right.