Don't care about which card is the turn card for abstraction.
When calculating equity distributions, the only important partition is
between the private cards and public cards, not which card came down on
which street. This distinctions reduces the number of canonical turn
hands by a factor of 3 and will probably make it feasible to store the
equity distributions in memory.