Implement k-means++ initialization for clustering
k-means++ is an initialization algorithm for choosing initial means for
k-means clustering. This will hopefully dramatically reduce the number
of iterations required to cluster the flop and turn hands.

I'm using the emd crate which has been a pain to get to compile because
it requires ndarrays, leading to some very messy code.