Implement the u64 hand representation
There was a bug in deal_canonical() where it didn't return all possible
canonical hands with streets=true. This means there are more canonical
hands than I thought, and the canonical river hands can't fit in memory
with the current hand representation of Vec<Card>, mainly due to the 24
byte Vec overhead. If I want a lookup table for the river abstraction
instead of on-the-fly equity calculation (slow), then I need to be able
to fit all canonical river hands in memory.

The u64 representation is a bitwise hand representation that stores 1 card as
1 byte. Since it's a primitive type, I don't have to worry about
lifetime parameters and borrowing like I would for an array or &str.