Fix bug where utilities were always positive
Turns out Rust isn't all good and it doesn't let you use variables in
match expressions. But instead of telling you by throwing and error or
not compiling, the variables just act as catch-all cases like _. So
everything just goes to the first case. Now the utilities seem to be
coming out correct and it needs be run for real.