Adjust pot size on the first action
I'm saying that the pot size = BIG_BLIND on the first action, that way
the bet abstraction will give us several bet sizes to choose from, not
just calling. Also, I fixed a bug where the sampled translated action
would not be a legal call in the true game, since the bot was not
"translating back" the call amount.