Implement LightAbstraction using buffered reading
We want to be able to look up river hands in the card abstraction
without needing to load an 8GB HashMap into memory. One slow easy way of
doing this is just to linear search through the entire
river_abstraction.txt file line by line until you find the hand you
want, and return the bucket ID number that's written next to it.

I will definitely speed this up later by partitioning river_abstraction.txt
into ~100 different files, but this just allows me to play against the
bot for now, even if it pauses for 30 seconds sometimes on the river.