Change the CompactBlueprint to store mixed strategies
I previously thought we only needed to store one chance-sampled action
for each infoset, but actually we need the full mixed strategy, which
I'm storing as 4-byte floats. This representation only appears to be 50%
larger than the chance-sampled action, probably because most of the size
is from storing the infoset.